Cosmos hub plugin and refactor for plugins module loading (#653)
* poc plugin

* plugins with a go.mod

* Added cosmos-hub plugin

* cosmos-hub plugin working and added parsing from CLI params

* added tests, refactored code to make it cleaner

* added documentation

* added documentation

* added the missing interfaces for osmosis

---------
