+++ tools/rosetta/README.md
@@ -67,6 +67,14 @@ go install cosmossdk.io/tools/rosetta/cmd/rosetta
 
 Alternatively, for building from source, simply run `make rosetta`. The binary will be located in `tools/rosetta`.
 
+## Plugins
+
+To use Rosetta standalone over any chain, it is possible to easily achieve it by setting up required prefixes and registering zone specific interfaces through plugins.
+
+Each plugin is a minimalist implementation of `InitZone` and `RegisterInterfaces` which allow Rosetta to parse chain specific data. There is an example for cosmos-hub chain under `plugins/default/` folder
+
+The plugin is selected through the cli `--blockchain` flag
+
 ## Extensions
 
 There are two ways in which you can customize and extend the implementation with your custom settings.
