+++ tools/rosetta/codec.go
@@ -6,13 +6,6 @@ import (
 	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
 	authcodec "github.com/cosmos/cosmos-sdk/x/auth/types"
 	bankcodec "github.com/cosmos/cosmos-sdk/x/bank/types"
-
-	ibcclienttypes "github.com/cosmos/ibc-go/v7/modules/core/02-client/types"
-//	"github.com/cosmos/ibc-go/v7/modules/core/exported"
-	ibcLightClient "github.com/cosmos/ibc-go/v7/modules/light-clients/07-tendermint"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	txtypes "github.com/cosmos/cosmos-sdk/types/tx"
 )
 
 // MakeCodec generates the codec required to interact
@@ -26,13 +19,4 @@ func MakeCodec() (*codec.ProtoCodec, codectypes.InterfaceRegistry) {
 	cryptocodec.RegisterInterfaces(ir)
 
 	return cdc, ir
-}
-
-// RegisterInterfaces registers rosetta related implementations and interfaces.
-func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
-	ibcclienttypes.RegisterInterfaces(registry)
-	ibcLightClient.RegisterInterfaces(registry)
-	sdk.RegisterInterfaces(registry)
-	txtypes.RegisterInterfaces(registry)
-	cryptocodec.RegisterInterfaces(registry)
 }
