+++ tools/rosetta/plugins/default/Makefile
@@ -1,0 +1,4 @@
+#!/usr/bin/make -f
+
+plugin:
+	go build -buildmode=plugin -o main.so main.go
